requirements.txt
paho-mqtt
scikit-learn
pandas
numpy
matplotlib
